crewai-agent/requirements.txt
crewai 
python-dotenv
openai 
# to install this crewai tools, do pip install `crewai[tools]`
tavily-python